+++ test/test_helper.rb
@@ -28,6 +28,10 @@ if ActiveSupport::TestCase.respond_to?(:fixture_path=)
   ActiveSupport::TestCase.fixtures :all
 end
 
+class ActiveSupport::TestCase
+  include ActiveJob::TestHelper
+end
+
 require "minitest/mock"
 require "mocha/minitest"